refactor(auth): make stub key data generation reusable
